+++ pssh/clients/reader.py
@@ -28,12 +28,13 @@ from gevent.lock import RLock
 class ConcurrentRWBuffer(object):
     """Concurrent reader/writer of bytes for use from multiple greenlets.
 
-    Iterate on buffer object to read data, blocking greenlet if no data exists
+    Iterate on buffer object to read data, yielding greenlet if no data exists
     until self.eof has been set.
 
     Writers should ``eof.set()`` when finished writing data via ``write``.
     Readers can use ``read()`` to get any available data, or None.
     """
+    __slots__ = ('_buffer', '_read_pos', '_write_pos', 'eof', '_lock')
 
     def __init__(self):
         self._buffer = BytesIO()
@@ -54,7 +55,10 @@ class ConcurrentRWBuffer(object):
             self._write_pos += self._buffer.write(data)
 
     def read(self):
-        """Read available data, or return None"""
+        """Read available data, or return None
+
+        :rtype: bytes
+        """
         with self._lock:
             if self._write_pos == 0 or self._read_pos == self._write_pos:
                 return