+++ tests/native/test_parallel_client.py
@@ -265,6 +265,19 @@ class ParallelSSHClientTest(unittest.TestCase):
         self.assertIsInstance(output[0].exception,
                               Timeout)
 
+    def test_timeout_on_open_session(self):
+        timeout = 1
+        client = ParallelSSHClient([self.host], port=self.port,
+                                   pkey=self.user_key,
+                                   timeout=timeout,
+                                   num_retries=1)
+        def _session(timeout=1):
+            sleep(timeout+1)
+        joinall(client.connect_auth())
+        sleep(.01)
+        client._host_clients[(0, self.host)].open_session = _session
+        self.assertRaises(Timeout, client.run_command, self.cmd)
+
     def test_connection_timeout(self):
         client_timeout = .01
         host = 'fakehost.com'